feat: refactor REST Api
