+++ components/crypto/handlers/record/remove/index.js
@@ -5,3 +5,12 @@ exports.request = function(msg, $meta) {
         payload: msg
     };
 };
+
+exports.error = function(err, $meta) {
+    throw this.errors.crypto({
+        cause: err,
+        details: {
+            statusCode: 404
+        }
+    });
+};
