+++ components/script/methods/remove/index.js
@@ -1,9 +1,10 @@
 module.exports = ({type, db}) => {
     return async function remove(msg, $meta) {
-        await this.bus.importMethod('crypto.record.remove')({id: msg.id});
+        const {id} = msg;
+        await this.bus.importMethod('crypto.record.remove')({id});
         if (db) {
-            await this.bus.importMethod(`db/${$meta.method}`)({externalId: msg.id});
+            await this.bus.importMethod(`db/${$meta.method}`)({externalId: id});
         }
-        return msg;
+        return {id};
     };
 };
