+++ components/swagger/crud/definitions/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    uuid: require('./uuid'),
+    error: require('./error'),
+    sync: require('./sync')
+};
