+++ components/swagger/crud/definitions/sync.js
@@ -1,0 +1,4 @@
+module.exports = {
+    type: 'boolean',
+    example: false
+};
