+++ components/swagger/crud/definitions/uuid.js
@@ -1,0 +1,4 @@
+module.exports = {
+    type: 'string',
+    pattern: '^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$'
+};
