+++ components/swagger/crud/methods/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    add: require('./add'),
+    edit: require('./edit'),
+    update: require('./update'),
+    fetch: require('./fetch'),
+    get: require('./get'),
+    remove: require('./remove')
+};
