+++ components/swagger/crud/methods/remove.js
